ai_service/requirements.txt
# Phase 4 Dependencies

apscheduler[asyncio]==3.10.1
category_encoders==2.6.2
celery==5.3.4
fastapi==0.111.0
imbalanced-learn==0.13.0
joblib==1.4.2
kaleido==0.2.1
lightgbm==4.1.0
lime==0.2.0.1
loguru==0.7.2
mlflow==2.9.2
motor==3.3.2
networkx==3.2.1
numpy==1.26.4
optuna==3.4.0
pandas==2.2.2
plotly==5.17.0
pymongo==4.7.2
python-dotenv==1.0.1
psutil==5.9.5
pytz==2023.3
redis==5.0.1
scikit-learn==1.6.1
scikit-optimize==0.9.0
scipy==1.11.4
shap==0.42.1
uvicorn==0.30.1
xgboost==1.7.6
